Move product lists into products section
