+++ cubedash/_platform.py
@@ -11,4 +11,4 @@ bp = Blueprint("platform", __name__, url_prefix="/platform")
 @bp.route("/<platform_name>")
 def platforms_page(platform_name):
     # Legacy platform page. Redirect to list of timelines in about page.
-    return redirect(url_for("about_page"))
+    return redirect(url_for("product.products_page"))
