+++ cubedash/templates/about.html
@@ -34,7 +34,7 @@
         <div class="header-follow">
             Open Data Cube {{ datacube_version }}
             {# TODO: Links to Website, CMI Metadata, DEA, ODC, help? #}
-            <a href="{{ url_for('about_sheet') }}" class="badge">csv <i class="fa fa-file-excel-o" aria-hidden="true"></i></a>
+            <a href="{{ url_for('product.products_csv') }}" class="badge">csv <i class="fa fa-file-excel-o" aria-hidden="true"></i></a>
         </div>
     </div>
     <div class="panel overflowable">
